Upload 와 Download 분리
