--- src/main/java/com/se/se_file_server/http/api/file/FileApiController.java
@@ -1,92 +1,0 @@
-package com.se.se_file_server.http.api.file;
-
-import com.se.se_file_server.domain.entity.file.File;
-
-import com.se.se_file_server.domain.usecase.file.FileUploadDownloadUseCase;
-
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.Resource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-@RestController
-@RequestMapping(value = "file")
-public class FileApiController {
-
-  private static final Logger logger = LoggerFactory.getLogger(FileApiController.class);
-
-  @Autowired
-  private FileUploadDownloadUseCase fileUploadDownloadUseCase;
-
-  @GetMapping("/")
-  public String controllerMain() {
-    return "Hello~ File Upload Test.";
-  }
-
-  @PostMapping("/uploadFile")
-  public File uploadFile(@RequestParam("file") MultipartFile file) {
-    String savedFileName = fileUploadDownloadUseCase.storeFile(file);
-
-    String fileDownloadUri = ServletUriComponentsBuilder.fromCurrentContextPath()
-        .path("/file/")
-        .path("/downloadFile/")
-        .path(savedFileName)
-        .toUriString();
-    
-    // 일단 테스트로 다운로드 Uri 넘겨줌
-    savedFileName = fileDownloadUri;
-
-    return new File(1L, file.getOriginalFilename(), savedFileName, file.getContentType(), file.getSize());
-  }
-
-  @PostMapping("/uploadMultipleFiles")
-  public List<File> uploadMultipleFiles(@RequestParam("files") MultipartFile[] files){
-    return Arrays.asList(files)
-        .stream()
-        .map(file -> uploadFile(file))
-        .collect(Collectors.toList());
-  }
-
-  @GetMapping("/downloadFile/{fileName:.+}")
-  public ResponseEntity<Resource> downloadFile(@PathVariable String fileName, HttpServletRequest request){
-    // Load file as Resource
-    Resource resource = fileUploadDownloadUseCase.loadFileAsResource(fileName);
-
-    // Try to determine file's content type
-    String contentType = null;
-    try {
-      contentType = request.getServletContext().getMimeType(resource.getFile().getAbsolutePath());
-    } catch (IOException ex) {
-      logger.info("Could not determine file type.");
-    }
-
-    // Fallback to the default content type if type could not be determined
-    if(contentType == null) {
-      contentType = "application/octet-stream";
-    }
-
-    return ResponseEntity.ok()
-        .contentType(MediaType.parseMediaType(contentType))
-        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename() + "\"")
-        .body(resource);
-  }
-
-}
